[Core] Rename errors to errorCodes
The core errors file only contains errorcodes.
Rename to more appropriate name, and prevent inadvertent confusion with
the C errors file.
