+++ src/openassetio-core/include/openassetio/errors/BatchElementError.hpp
@@ -7,7 +7,7 @@
 #include <string>
 #include <utility>
 
-#include <openassetio/errors/errors.h>
+#include <openassetio/errors/errorCodes.h>
 #include <openassetio/export.h>
 #include <openassetio/typedefs.hpp>
