+++ src/openassetio-core/include/openassetio/errors/errorCodes.h
